Update add controller

+++ Data/Dto/OrderDTO.cs
@@ -2,22 +2,11 @@ namespace webapp.Dto
 {
     public class OrderDto
     {
-        // public class OrderDetailsDto
-        // {
-            public string ItemID { get; set; }
-
-            public int Quantity { get; set; }
-
-            public float Price { get; set; }
-            
-            public int Amount { get; set; }
-        // }
-
-        // public class OrderMasterDto
-        // {
-            public string CustomerID { get; set; }
-
-            public float TotalAmount { get; set; }
-        // }
+        public string ItemID { get; set; }
+        public int Quantity { get; set; }
+        public float Price { get; set; }
+        public decimal Amount { get; set; }
+        public string CustomerID { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
